Use store to communicate between components
