+++ frontend/src/store.js
@@ -5,10 +5,14 @@ Vue.use(Vuex)
 
 export default new Vuex.Store({
   state: {
-
+    token: ""
+  }, getters: {
+    token: state => state.token,
   },
   mutations: {
-
+    setToken(state, token) {
+      state.token = token;
+    }
   },
   actions: {
